refactor(pekofy): moved Pekogacha to pekofy.py and added a couple more messages
